Add image source extraction to productSearchResult
The changes introduce the extraction of an image source URL from the `bestMatchingItem` object. Specifically, the code now retrieves the first element from the `cse_thumbnail` array within the `Pagemap` property of `bestMatchingItem`. This element is then used to set the `ImageSource` property of the `productSearchResult` object. If the image source is not found, it defaults to an empty string.